requirements.txt
django
djangorestframework
django-environ
drf-yasg
tinydb